+++ packages/hermes-image/nixos-base-module.nix
@@ -1,6 +1,7 @@
 { ... }:
 {
   _module.args.includeRepoContent = false;
+  _module.args.includeManagedRuntime = false;
   imports = [
     ./nixos-module.nix
   ];