+++ packages/hermes-image/nixos-final-module.nix
@@ -1,6 +1,7 @@
 { ... }:
 {
   _module.args.includeRepoContent = true;
+  _module.args.includeManagedRuntime = true;
   imports = [
     ./nixos-module.nix
   ];